:sparkles: 댓글수정 기능 구현 완료

+++ src/main/kotlin/com/teamsparta/exhibitionnewsfeed/domain/newsfeed/comment/repository/CommentRepository.kt
@@ -3,4 +3,4 @@ package com.teamsparta.exhibitionnewsfeed.domain.newsfeed.comment.repository
 import com.teamsparta.exhibitionnewsfeed.domain.newsfeed.comment.model.Comment
 import org.springframework.data.jpa.repository.JpaRepository
 
-interface CommentRepository : JpaRepository<Comment, Int>
+interface CommentRepository : JpaRepository<Comment, Long>
